Lionengine 8.0.0

+++ lionengine-core/src/main/java/com/b3dgs/lionengine/UtilFile.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2013-2014 Byron 3D Games Studio (www.b3dgs.com) Pierre-Alexandre (contact@b3dgs.com)
+ * Copyright (C) 2013-2015 Byron 3D Games Studio (www.b3dgs.com) Pierre-Alexandre (contact@b3dgs.com)
  * 
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -18,15 +18,20 @@
 package com.b3dgs.lionengine;
 
 import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Locale;
 
-import com.b3dgs.lionengine.core.Core;
+import com.b3dgs.lionengine.core.Medias;
 import com.b3dgs.lionengine.core.Verbose;
 
 /**
  * Static functions giving informations related to files and directory.
+ * <p>
+ * This class is Thread-Safe.
+ * </p>
  * 
  * @author Pierre-Alexandre (contact@b3dgs.com)
  */
@@ -165,7 +170,7 @@ public final class UtilFile
      */
     public static String getFilenameFromPath(String path)
     {
-        final int i = path.lastIndexOf(Core.MEDIA.getSeparator());
+        final int i = path.lastIndexOf(Medias.getSeparator());
         return path.substring(i + 1, path.length());
     }
 
@@ -178,12 +183,12 @@ public final class UtilFile
     public static String[] getDirsList(String path)
     {
         final File file = new File(path);
-        if (!file.exists())
+        final File[] files = file.listFiles();
+        if (files == null)
         {
             return new String[0];
         }
 
-        final File[] files = file.listFiles();
         int numberOfDirs = 0;
         for (final File file2 : files)
         {
@@ -192,7 +197,6 @@ public final class UtilFile
                 numberOfDirs++;
             }
         }
-
         final String[] dirsList = new String[numberOfDirs];
         for (int i = 0, id = 0; i < files.length; i++)
         {
@@ -215,12 +219,11 @@ public final class UtilFile
     public static String[] getFilesList(String path)
     {
         final File file = new File(path);
-        if (!file.exists())
+        final File[] files = file.listFiles();
+        if (files == null)
         {
             return new String[0];
         }
-        final File[] files = file.listFiles();
-        String[] filesList;
 
         int numberOfFiles = 0;
         for (final File file2 : files)
@@ -230,8 +233,7 @@ public final class UtilFile
                 numberOfFiles++;
             }
         }
-
-        filesList = new String[numberOfFiles];
+        final String[] filesList = new String[numberOfFiles];
         for (int i = 0, id = 0; i < files.length; i++)
         {
             if (files[i].isFile())
@@ -240,7 +242,6 @@ public final class UtilFile
                 id++;
             }
         }
-
         return filesList;
     }
 
@@ -295,13 +296,20 @@ public final class UtilFile
         if (directory.isDirectory())
         {
             final String[] children = directory.list();
-            for (final String element : children)
+            if (children != null)
             {
-                deleteDirectory(new File(directory, element));
+                for (final String element : children)
+                {
+                    deleteDirectory(new File(directory, element));
+                }
             }
-            if (!directory.delete())
+            try
             {
-                Verbose.warning(UtilFile.class, "deleteDirectory", "Directory not deleted: " + directory);
+                Files.delete(directory.toPath());
+            }
+            catch (final IOException exception)
+            {
+                Verbose.exception(UtilFile.class, "deleteDirectory", exception, "Directory not deleted: " + directory);
             }
         }
         else if (directory.isFile())
@@ -319,10 +327,13 @@ public final class UtilFile
     public static void deleteFile(File file)
     {
         Check.notNull(file);
-
-        if (file.isFile() && !file.delete())
+        try
         {
-            Verbose.warning(UtilFile.class, "deleteDir", "File not deleted: " + file);
+            Files.delete(file.toPath());
+        }
+        catch (final IOException exception)
+        {
+            Verbose.warning(UtilFile.class, "deleteFile", "File not deleted: " + file);
         }
     }
 
@@ -339,15 +350,18 @@ public final class UtilFile
         if (file.exists())
         {
             final File[] files = file.listFiles();
-            for (final File content : files)
+            if (files != null)
             {
-                if (content.isDirectory())
+                for (final File content : files)
                 {
-                    getFilesByExtensionRecursive(filesList, content.getPath(), extension);
-                }
-                if (content.isFile() && extension.equals(getExtension(content)))
-                {
-                    filesList.add(content);
+                    if (content.isDirectory())
+                    {
+                        getFilesByExtensionRecursive(filesList, content.getPath(), extension);
+                    }
+                    if (content.isFile() && extension.equals(getExtension(content)))
+                    {
+                        filesList.add(content);
+                    }
                 }
             }
         }
@@ -362,15 +376,19 @@ public final class UtilFile
      */
     private static void getFilesByNameRecursive(Collection<File> filesList, File path, String name)
     {
-        for (final File file : path.listFiles())
+        final File[] files = path.listFiles();
+        if (files != null)
         {
-            if (file.isFile() && file.getName().equals(name))
+            for (final File file : files)
             {
-                filesList.add(file);
-            }
-            else if (file.isDirectory())
-            {
-                getFilesByNameRecursive(filesList, file, name);
+                if (file.isFile() && file.getName().equals(name))
+                {
+                    filesList.add(file);
+                }
+                else if (file.isDirectory())
+                {
+                    getFilesByNameRecursive(filesList, file, name);
+                }
             }
         }
     }
